Removing code that will never be used.
