+++ src/fatfs.c
@@ -36,21 +36,18 @@
 uint8_t retSD;    /* Return value for SD */
 char SD_Path[4];  /* SD logical drive path */
 
-/* USER CODE BEGIN Variables */
 
-/* USER CODE END Variables */    
-
-void MX_FATFS_Init(void) {
+void MX_FATFS_Init() {
   /*## FatFS: Link the SD driver ###########################*/
   retSD = FATFS_LinkDriver(&SD_Driver, SD_Path);
 
   /* USER CODE BEGIN Init */
-  /* additional user code for init */     
+  /* additional user code for init */
   /* USER CODE END Init */
 }
 
 /**
-  * @brief  Gets Time from RTC 
+  * @brief  Gets Time from RTC
   * @param  None
   * @retval Time in DWORD
   */
@@ -58,4 +55,3 @@ DWORD get_fattime(void) {
   /* USER CODE BEGIN get_fattime */
   return 0;
 }
-
